pyproject.toml
[project]
name = "hydrohomie"
version = "0.1.0"
description = "Reminds people to drink their water"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "atproto>=0.0.65",
    "pyyaml>=6.0.3",
]
